move CardPresentPaymentServiceScreenshotMock
